[SY] feat[#71]: 도서, 아이템 이미지 url 수정 및 uploadimage api 컴포넌트화
